pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Hansotto Kristiansen", email = "hansotto.kristiansen@gmail.com"},
    {name = "", email = "hansotto.kristiansen@gmail.com"},
]
dependencies = [
    "pydantic>=2.8.2",
    "pydantic-settings>=2.3.4",
    "typing-extensions>=4.12.2",
    "python-dotenv>=1.0.1",
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "langchain>=0.2.16",
    "langchain-community>=0.2.16",
    "azure-cosmos>=4.7.0",
    "langchain-openai>=0.2.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.11",
    "pypdf>=5.0.0",
]

requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.4.3",
]
lint = [
    "black>=23.10.1", 
]



